Revert "début changement archi pour Chrominos ; ne compile pas"
début changement archi pour Chrominos ; ne compile pas

+++ Data/Models/ChrominoInGame.cs
@@ -12,7 +12,7 @@ namespace Data.Models
     public class ChrominoInGame
     {
         public int Id { get; set; }
-        public int ChrominoId { get; set; }
+        public byte ChrominoId { get; set; }
         public int GameId { get; set; }
         public Orientation Orientation { get; set; }
         public int XPosition { get; set; }
